BD Tool BE/requirements.txt
fastapi==0.111.0
fuzzywuzzy==0.18.0
jose==1.0.0
pandas==2.0.3
passlib==1.7.4
pydantic==1.10.15
pymongo==3.13.0
python-dotenv==1.0.1
python_jose==3.3.0
uvicorn==0.29.0
